UX improvement
I made the opening UX better. Just small things while I think on how to do the data while I can't see it directly, as the bitwise things aren't jiving to give me sensible data. But I will persevere and form the data as need so all we need is the parsing so tomorrow can solely be about parsing bits