0.1.6 release candidate
This is the first community found bug which is pretty obvious to me now. I hadn't set up the release settings and only had the debug building correctly. I had to disable some optimization cause it was breaking the linker however now it builds as expected. Additionally I added logging for a handful of members reporting a lack of connection to the program while the official driver is functioning. That will become the next community update once resolved. however I need more testing from the community